% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/central_bank.R
\name{central_bank}
\alias{central_bank}
\title{central_bank generator}
\usage{
central_bank(
  pays = NA,
  bonds = NA,
  numeraire = NA,
  depot_banque_centrale = NA,
  inflation_cible = NA,
  overnight_rate_cible = inflation_cible * 2,
  fourchette = function(overnight_cible, depot_banque_centrale) overnight_cible +
    max(-0.0025, min(0.0025, 0.0025 - depot_banque_centrale * 0.005/100)),
  base_monetaire = NA
)
}
\arguments{
\item{pays}{un character : pays.}

\item{bonds}{une liste : bonds possédés par la banque centrale (actif).}

\item{numeraire}{un numeric : quantité d'argent en circulation hors banques
(passif).}

\item{depot_banque_centrale}{un numeric : réserves que les banques à charte
prêtent à la banque centrale au taux du financement à un jour (passif si
positif).}

\item{inflation_cible}{un numeric : taux d'inflation ciblé.}

\item{overnight_rate_cible}{un numeric : taux d'intérêt à un jour ciblé pour
atteindre la cible d'inflation.}

\item{fourchette}{une fonction : équation de la fourchette opérationelle
pour le taux de financement à un jour.}

\item{base_monetaire}{un numeric : base monétaire (passif).

Si le canada veut atteindre 2\% d'inflation (taux constant) et que le taux
d'inflation courant est à 5\%, il faut monter les taux d'intérêt. La banque
centrale va augmenter le taux cible du financement à un jour à supposons,
10\%. Elle va donc se mettre à échanger davantage avec les banques à charte
avec des opérations d'open market changeant la quantité d'argent disponible
sur le marché.}
}
\value{
objet de classe "central_bank".
}
\description{
central_bank generator
}
\examples{
central_bank(
  "Canada",
  list(
    bond(1000, 0.03, 30, 18),
    bond(2000, 0.04, 15, 4)
  ),
  100000,
  20000,
  0.02
)
}
