% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{CPG}
\alias{CPG}
\title{CPG generator}
\usage{
CPG(
  face_value = 0,
  interest = 0,
  expiration = 0,
  time_left = 0,
  cash_flow = NA
)
}
\arguments{
\item{face_value}{un numeric : valeur nominale du CPG.}

\item{interest}{un numeric : taux d'intérêt du CPG.}

\item{expiration}{un numeric : durée totale du CPG.}

\item{time_left}{un numeric : durée restante du CPG.}

\item{cash_flow}{un numeric : montant qui sera versé à la fin de l'année.}
}
\value{
objet de classe "CPG".
}
\description{
CPG generator
}
\examples{
CPG(
  1000,
  0.03,
  30,
  18
)
}
