% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{dette}
\alias{dette}
\title{dette generator}
\usage{
dette(
  face_value = 0,
  interest = 0,
  expiration = 0,
  time_left = 0,
  cash_flow = NA
)
}
\arguments{
\item{face_value}{un numeric : valeur nominale du dette.}

\item{interest}{un numeric : taux d'intérêt du dette.}

\item{expiration}{un numeric : durée totale du dette.}

\item{time_left}{un numeric : durée restante du dette.}

\item{cash_flow}{un numeric : montant qui sera versé à la fin de l'année.}
}
\value{
objet de classe "dette".
}
\description{
dette generator
}
\examples{
dette(
  1000,
  0.03,
  30,
  18
)
}
