% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devise.R
\name{devise}
\alias{devise}
\title{devise generator}
\usage{
devise(nom = NA, taux_change = NA)
}
\arguments{
\item{nom}{un character : nom de la devise.}

\item{taux_change}{un numeric : taux de change avec la devise par défaut : le
US.}
}
\value{
objet de classe "devise".
}
\description{
Pas besoin de définir le dollar US.
}
\examples{
devise("Yen", 90.3)
}
