\name{economics-package}
\alias{economics-package}
\alias{economics}
\docType{package}
\title{
\packageTitle{economics}
}
\description{
\packageDescription{economics}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{economics}
\packageIndices{economics}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{economics}

Maintainer: \packageMaintainer{economics}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
