% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/government.R
\name{government}
\alias{government}
\title{government generator}
\usage{
government(
  pays = NA,
  taxes = NA,
  impot = NA,
  bonds = NA,
  dette = NA,
  min_payment = NA
)
}
\arguments{
\item{pays}{un character : nom du pays associé.}

\item{taxes}{un numeric : taxes de ventes.}

\item{impot}{une fonction : impot selon les revenus.}

\item{bonds}{une liste : tous les bonds émis par le gouvernement.}

\item{dette}{un numeric : dette publique.}

\item{min_payment}{un numeric : intérêt minimal à rembourser à chaque année.}
}
\value{
objet de classe "government".
}
\description{
government generator
}
\examples{
government("Canada", 0.15, function(revenu) 0.3 * revenu, list())
government("Canada", 0.15, function(revenu) 0.3 * revenu, list(bond(1000, 0.03, 30, 18), bond(10000, 0.02, 15, 1)))
}
