% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitant.R
\name{habitant}
\alias{habitant}
\title{habitant generator}
\usage{
habitant(
  id = NA,
  pays = NA,
  numeraire = NA,
  compte_cheque = NA,
  CPG = NA,
  credit = NA,
  dette = NA,
  bonds = NA,
  bonheur = NA,
  biens = list(),
  formule_bonheur = function(biens) {     NA }
)
}
\arguments{
\item{id}{un entier : id de l'habitant}

\item{pays}{un character : nom du pays de provenance.}

\item{credit}{un numeric : crédit total en US dans le compte de l'habitant.}

\item{dette}{un numeric : dette totale en US dans le compte de l'habitant.}

\item{bonds}{une liste de \code{bonds} possédés par l'habitant.}

\item{bonheur}{un numeric : niveau de bonheur de l'habitant.}

\item{biens}{une liste de biens possédés par l'habitant.}

\item{formule_bonheur}{une formule : formule qui calcule le bonheur de
l'habitant selon les biens qu'il possède.}

\item{argent}{un numeric : argent total en US dans le compte de l'habitant.}
}
\value{
objet de classe "habitant".
}
\description{
habitant generator
}
\examples{
habitant(
  7,
  "Canada",
  5000,
  20000,
  40000,
  0,
  0,
  list(
    bond(
      1000,
      0.04,
      30
    ),
    bond(
      2000,
      0.03,
      15
    )
  ),
  73,
  list(
    eau = 2,
    bouffe = 30
  ),
  function(biens) {
    biens$bouffe * 0.1 + biens$eau * 2
  }
)
}
