% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pays.R
\name{pays}
\alias{pays}
\title{pays generator}
\usage{
pays(
  nom = NA,
  population = NA,
  devise = NA,
  GDP = NA,
  unemployment = NA,
  inflation = NA,
  central_bank = NA,
  government = NA
)
}
\arguments{
\item{nom}{un character : nom du pays.}

\item{population}{un integer : nombre d'habitants du pays.}

\item{devise}{un character : devise du pays.}

\item{GDP}{un numeric : GDP du pays.}

\item{unemployment}{un numeric : taux de chomage du pays.}

\item{inflation}{un numeric : inflation du pays.}

\item{central_bank}{un objet de classe "central_bank".}

\item{government}{un objet de classe "government".}
}
\value{

}
\description{
pays generator
}
