% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people.R
\name{people}
\alias{people}
\title{people generator}
\usage{
people(n = 1, pays = NA)
}
\arguments{
\item{n}{un integer : nombre d'habitants à générer aléatoirement.}

\item{pays}{un vecteur de character : noms des pays existants.}
}
\value{
liste d'habitants.
}
\description{
people generator
}
\examples{
people(3)

peuple <- people(3)
numeraire(peuple)
M1(peuple)
M2(peuple)
}
